requirements/requirements.txt
Django>=2.2

django-db-mutex>=2.0.0
django-entity>=5.0.0
django-entity-event>=2.0.0
ambition-django-uuidfield>=0.5.0
ambition-utils>=2.2.0

beautifulsoup4>=4.3.2
